handle json or stream on auto response

+++ frontend/src/chat-gpt.ts
@@ -7,7 +7,7 @@ import { chatGptStyles } from './chat-gpt.css.js';
 import { githubDarkDimmed } from './github-dark-dimmed.css.js';
 import { loadingIcon } from './loading-icon.js';
 import { store } from './store.js';
-import { ChatHistory, Engine, Thread } from './types.js';
+import { Thread } from './types.js';
 const renderer = {
   image(href = '', title = 'image', text = 'image') {
     return `
@@ -46,21 +46,16 @@ export class ChatGPT extends LitElement {
   `;
   @query('textarea') textareaEl: HTMLTextAreaElement;
   @state() loading = false;
-  @state() history: ChatHistory = [];
   @state() miniPreviewImageURL = '';
-  @state() system_message: string;
-  @state() engine: Engine;
-  @state() include_context: boolean;
+  @state() thread: Thread;
   connectedCallback() {
     super.connectedCallback();
     store.subscribe<Thread>('activeThread', (thread) => {
-      this.history = thread.history;
-      this.engine = thread.selected_engine;
-      this.include_context = thread.include_context;
+      this.thread = thread;
       this.requestUpdate();
     });
   }
-  async performPostRequest(endpoint: string, body: any): Promise<any> {
+  async performPostRequest(endpoint: string, body: Thread | FormData): Promise<any> {
     this.loading = true;
     let headers = {};
     if (body instanceof FormData) {
@@ -81,9 +76,8 @@ export class ChatGPT extends LitElement {
     return response;
   }
 
-  async submit(e: Event) {
-    const engine = this.engine;
-    switch (engine) {
+  async submit() {
+    switch (this.thread.selected_engine) {
       case 'gpt-4-1106-preview':
       case 'gpt-4':
       case 'gpt-3.5-turbo':
@@ -102,6 +96,9 @@ export class ChatGPT extends LitElement {
       case 'vertex':
         await this.runVertexReq();
         break;
+      case 'auto':
+        await this.runAutoReq();
+        break;
       default:
         console.log('invalid engine');
         break;
@@ -109,35 +106,18 @@ export class ChatGPT extends LitElement {
   }
 
   async runChatReq() {
-    const engine = this.engine;
     const element = this.textareaEl;
     const prompt = element.value;
-    const includeContext = this.include_context;
     element.value = '';
     await this.addToHistory('user', prompt);
 
-    const reqBody = {
-      ...(includeContext
-        ? {
-            messages: [
-              { role: 'system', content: this.system_message },
-              ...this.history.map((item) => ({ role: item.role, content: item.content }))
-            ]
-          }
-        : {
-            messages: [
-              { role: 'system', content: this.system_message },
-              { role: 'user', content: prompt }
-            ]
-          }),
-      engine: engine
-    };
+    const reqBody: Thread = this.thread;
     try {
       const response = await this.performPostRequest('/', reqBody);
 
       const reader = response.body.getReader();
       let message = '';
-      const nextIndex = this.history.length;
+      const nextIndex = this.thread.history.length;
       this.addToHistory('assistant', '');
 
       while (true) {
@@ -156,133 +136,145 @@ export class ChatGPT extends LitElement {
   }
 
   async runVertexReq() {
-    const element = this.textareaEl;
-    const prompt = element.value;
-    element.value = '';
-    const includeContext = this.include_context;
-    this.addToHistory('user', prompt);
-
-    const requestBody = {
-      ...(includeContext
-        ? {
-            messages: [
-              ...this.history.map((item) => ({
-                role: item.role === 'assistant' ? 'bot' : 'user',
-                content: item.content
-              }))
-            ]
-          }
-        : {
-            messages: [{ role: 'user', content: prompt }]
-          })
-    };
-
-    try {
-      const response = await this.performPostRequest('/vertex', requestBody);
-      const json = await response.json();
-
-      if (!json.content) {
-        throw new Error('No content in response');
-      }
-      // Assuming `json.content` is the part of the response you want to display
-      this.addToHistory('assistant', json.content);
-    } catch (err: any) {
-      console.error(err);
-      this.addToHistory('assistant', `Vertex AI error: ${err.message}. Try again.`);
-    } finally {
-      this.loading = false;
-    }
+    // const element = this.textareaEl;
+    // const prompt = element.value;
+    // element.value = '';
+    // await this.addToHistory('user', prompt);
+    // const requestBody = this.thread;
+    // try {
+    //   const response = await this.performPostRequest('/vertex', requestBody);
+    //   const json = await response.json();
+    //   if (!json.content) {
+    //     throw new Error('No content in response');
+    //   }
+    //   // Assuming `json.content` is the part of the response you want to display
+    //   this.addToHistory('assistant', json.content);
+    // } catch (err: any) {
+    //   console.error(err);
+    //   this.addToHistory('assistant', `Vertex AI error: ${err.message}. Try again.`);
+    // } finally {
+    //   this.loading = false;
+    // }
   }
 
   async runAutoReq() {
-    // auto request first gets the prompt content
-    // it checks for the presence of an image
-    // it then posts to auto with either applicaton/json or multipart/form-data
-    // it also checks for includecontext
-    // it will then await the response
-    // the response is standardized to be some sort of content for the assistant messag
+    const element = this.textareaEl;
+    const prompt = element.value;
+    element.value = '';
+    await this.addToHistory('user', prompt);
+
+    const reqBody: Thread = this.thread;
+    try {
+      const response = await this.performPostRequest('/auto', reqBody);
+
+      if (response.headers.get('content-type')?.includes('application/json')) {
+        const json = await response.json();
+        if (!json.content) {
+          throw new Error('No content in response');
+        }
+        // Assuming `json.content` is the part of the response you want to display
+        this.addToHistory('assistant', json);
+      } else if (response.headers.get('content-type')?.includes('text/event-stream')) {
+        const reader = response.body.getReader();
+        let message = '';
+        const nextIndex = this.thread.history.length;
+        this.addToHistory('assistant', '');
+
+        while (true) {
+          const { done, value } = await reader.read();
+          if (done) break;
+          message = new TextDecoder().decode(value);
+          this.updateAssistantResponse(nextIndex, message);
+          this.loading = false;
+          this.shadowRoot.querySelector('.history-outer').scrollTop = 0;
+        }
+      }
+    } catch (err: any) {
+      console.error(err);
+      this.addToHistory('assistant', `auto api error: ${err.message}. Try again.`);
+    } finally {
+      this.loading = false;
+    }
   }
 
   async runVisionReq() {
-    const element = this.textareaEl;
-    const prompt = element.value;
-    element.value = '';
-    this.miniPreviewImageURL = '';
-    try {
-      const form = new FormData();
-      const file = this.shadowRoot.querySelector<HTMLInputElement>('input[type=file]').files[0];
-      form.append('file', file);
-      this.addToHistory('user', `![image](${URL.createObjectURL(file)})`);
-      const response = await this.performPostRequest('/vision', form);
-      if (response.status !== 200) {
-        throw new Error('http error. try again');
-      }
-      const json = await response.json();
-      if (!json.content) {
-        throw new Error('no content in response');
-      }
-      this.addToHistory('assistant', json.content);
-    } catch (err) {
-      this.addToHistory('assistant', `http error.${err} try again`);
-    } finally {
-      this.loading = false;
-    }
+    // const element = this.textareaEl;
+    // const prompt = element.value;
+    // element.value = '';
+    // this.miniPreviewImageURL = '';
+    // try {
+    //   const form = new FormData();
+    //   const file = this.shadowRoot.querySelector<HTMLInputElement>('input[type=file]').files[0];
+    //   form.append('file', file);
+    //   this.addToHistory('user', `![image](${URL.createObjectURL(file)})`);
+    //   const response = await this.performPostRequest('/vision', form);
+    //   if (response.status !== 200) {
+    //     throw new Error('http error. try again');
+    //   }
+    //   const json = await response.json();
+    //   if (!json.content) {
+    //     throw new Error('no content in response');
+    //   }
+    //   this.addToHistory('assistant', json.content);
+    // } catch (err) {
+    //   this.addToHistory('assistant', `http error.${err} try again`);
+    // } finally {
+    //   this.loading = false;
+    // }
   }
 
   async runImageReq() {
-    const engine = this.engine;
-    const element = this.textareaEl;
-    const prompt = element.value;
-    element.value = '';
-    this.addToHistory('user', prompt);
-
-    const requestBody = {
-      model: engine,
-      prompt: prompt,
-      n: 1,
-      size: '1024x1024'
-    };
-    try {
-      const response = await this.performPostRequest('/image', requestBody);
-      const imgResp = (await response.json()) as { url: string };
-      const imgMarkdown = `![image](${imgResp.url})`;
-      this.addToHistory('assistant', imgMarkdown);
-    } catch (err) {
-      this.addToHistory('assistant', 'http error. try again');
-    } finally {
-      this.loading = false;
-    }
+    // const engine = this.thread.selected_engine;
+    // const element = this.textareaEl;
+    // const prompt = element.value;
+    // element.value = '';
+    // await this.addToHistory('user', prompt);
+    // const requestBody = {
+    //   model: engine,
+    //   prompt: prompt,
+    //   n: 1,
+    //   size: '1024x1024'
+    // };
+    // try {
+    //   const response = await this.performPostRequest('/image', requestBody);
+    //   const imgResp = (await response.json()) as { url: string };
+    //   const imgMarkdown = `![image](${imgResp.url})`;
+    //   this.addToHistory('assistant', imgMarkdown);
+    // } catch (err) {
+    //   this.addToHistory('assistant', 'http error. try again');
+    // } finally {
+    //   this.loading = false;
+    // }
   }
 
   async runTtsReq() {
-    const element = this.textareaEl;
-    const prompt = element.value;
-    element.value = '';
-    this.addToHistory('user', prompt);
-    const requestBody = {
-      model: 'tts-1',
-      input: prompt,
-      voice: 'nova'
-    };
-
-    try {
-      const response = await this.performPostRequest('/tts', requestBody);
-      if (response.ok) {
-        const blob = await response.blob();
-        const audioUrl = URL.createObjectURL(blob);
-        this.addToHistory(
-          'assistant',
-          `Audio response received. Click to play.`,
-          `<audio controls src="${audioUrl}"></audio>`
-        );
-      } else {
-        throw new Error('Non-OK response from TTS endpoint');
-      }
-    } catch (err) {
-      this.addToHistory('assistant', 'TTS error. Try again.');
-    } finally {
-      this.loading = false;
-    }
+    // const element = this.textareaEl;
+    // const prompt = element.value;
+    // element.value = '';
+    // this.addToHistory('user', prompt);
+    // const requestBody = {
+    //   model: 'tts-1',
+    //   input: prompt,
+    //   voice: 'nova'
+    // };
+    // try {
+    //   const response = await this.performPostRequest('/tts', requestBody);
+    //   if (response.ok) {
+    //     const blob = await response.blob();
+    //     const audioUrl = URL.createObjectURL(blob);
+    //     this.addToHistory(
+    //       'assistant',
+    //       `Audio response received. Click to play.`,
+    //       `<audio controls src="${audioUrl}"></audio>`
+    //     );
+    //   } else {
+    //     throw new Error('Non-OK response from TTS endpoint');
+    //   }
+    // } catch (err) {
+    //   this.addToHistory('assistant', 'TTS error. Try again.');
+    // } finally {
+    //   this.loading = false;
+    // }
   }
 
   async addToHistory(role: 'user' | 'assistant', content: string, custom?: string) {
@@ -294,14 +286,15 @@ export class ChatGPT extends LitElement {
     store.addToMessageContent(newContent, index);
   }
   deleteItem(index: number) {
-    const reversedIndex = this.history.length - index - 1;
+    const reversedIndex = this.thread.history.length - index - 1;
     store.deleteChatHistoryItem(reversedIndex);
   }
   render() {
+    if (!this.thread) return nothing;
     return html`
       <div class="history-outer">
         ${loadingIcon(this.loading)}
-        ${[...this.history].reverse().map(
+        ${[...this.thread.history].reverse().map(
           (item, i) => html`
             <p class="history ${item.role}">
               <button class="delete" @click=${(e) => this.deleteItem(i)}>x</button>
@@ -313,13 +306,13 @@ export class ChatGPT extends LitElement {
       <div class="inputs-outer">
         <div class="inputs-inner">
           <textarea
-            ?disabled=${this.engine === 'gpt-4-vision-preview'}
+            ?disabled=${this.thread.selected_engine === 'gpt-4-vision-preview'}
             title="Enter to send. Shift+Enter for new line."
             @keydown=${(e) => {
               if (e.key === 'Enter' && e.shiftKey) return;
               if (e.key === 'Enter') {
                 e.preventDefault();
-                this.submit(e);
+                this.submit();
                 return;
               }
             }}></textarea>
@@ -327,7 +320,7 @@ export class ChatGPT extends LitElement {
             ${this.miniPreviewImageURL
               ? html` <div class="mini-preview"><img src="${this.miniPreviewImageURL}" /></div>`
               : nothing}
-            ${this.engine?.includes('vision')
+            ${this.thread.selected_engine?.includes('vision')
               ? html`
                   <input
                     type="file"
